Create CheckpointInitialPosition on SubscriptionWithCheckpointOptions
Using 3 record types, this commit allows a subscription to define its
starting point.

It can be Beginning, End or From a specific position.Create CheckpointInitialPosition on SubscriptionWithCheckpointOptions

Using 3 record types, this commit allows a subscription to define its
starting point.

It can be Beginning, End or From a specific position.